requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.1
html5lib>=1.1
beautifulsoup4>=4.12.2
selenium>=4.15.2
Pillow>=10.1.0
